ngpu: change resources locations and bindings type to uint32_t
Locations and bindings are uint32_t on Vulkan, Metal, DirectX, WGPU and in some
part of the OpenGL API.

Fixes -Wsign-conversion warnings.

+++ libnopegl/src/ngpu/opengl/program_gl_utils.c
@@ -83,7 +83,7 @@ int ngpu_program_gl_set_locations_and_bindings(struct ngpu_program *s,
         const struct ngpu_bindgroup_layout_entry *entry = &layout_desc.textures[i];
         const char *texture_name = ngpu_pgcraft_get_symbol_name(crafter, entry->id);
         const GLint location = gl->funcs.GetUniformLocation(s_priv->id, texture_name);
-        gl->funcs.Uniform1i(location, entry->binding);
+        gl->funcs.Uniform1i(location, (GLint)entry->binding);
     }
 
     return 0;
